Use OTP provided CA trust store if available
This patch is especially useful on Windows which does not have any of
the known locations but ships with a CA nonetheless. (OTP fetches it
using NIF.)

Additionally, if `ELIXIR_MAKE_CACERT` is set it will be used even if the
file it is pointing to does not exist in which case the user will get an
error.